Fix the countdown timer
I also removed the logic around the TWAP oracles. I realize that there
is a trade-off here, and I understand why it was added in the first
place, but I think the UX of seeing an inaccurate time (e.g., at the
moment, seeing that the proposal finalizes in 5 days since we haven't
had any TWAP observations) is worse than the UX of trying to finalize
but being unable to.